adjust everything

+++ src/services/post.service.ts
@@ -7,10 +7,13 @@ import { DatetimeService } from './datetime.service';
 import { CategoryService } from './category.service';
 import { Category } from 'src/entity/category';
 import { TagService } from './tag.service';
+import { UserService } from './user.service';
+import { UserDto } from 'src/validation/user.dto';
+import { ImageService } from './image.service';
 @Injectable()
 export class PostService {
     constructor(@InjectRepository(Posts) private postRepository: Repository<Posts>, private dateTime: DatetimeService,
-                private categoryService: CategoryService, private tagService: TagService){}
+                private categoryService: CategoryService, private tagService: TagService, private userService: UserService){}
 
     async Add(post: PostDto): Promise<any>{
         const publishedDate = this.dateTime.GetDateTimeString();
@@ -26,12 +29,29 @@ export class PostService {
                 tags.push(tag);
             }
         }
-        await this.postRepository.save({title: post.title, content: post.content, 
+        const newPost = await this.postRepository.save({title: post.title, content: post.content, 
                                         likedCount: 0, sharedCount: 0, savedCount: 0, 
                                         commentCount: 0, publishedDate:publishedDate,
                                         user:{id: post.userId}, access: {id: post.accessId}, 
                                         categories: categories, tags: tags});
-        return{message: "Post successfully !"};
+        const user = await this.userService.FindOne(undefined, undefined , post.userId);
+        const updatedUser: UserDto = {
+            publishedPostCount: user.postPublishedCount += 1,
+            firstName: user.firstName, 
+            lastName: user.lastName, 
+            phoneNum: user.phoneNum, 
+            password: user.password,
+            email: user.email, 
+            bio: user.bio,
+            profilePicturePath: user.profilePicturePath,
+            dateOfBirth: user.dateOfBirth,
+            username: user.username,
+            confirmPassword: user.password,
+            gender: user.gender,
+            roleId: user.role.id
+        }
+        await this.userService.UpdateUserInfor(user.id, updatedUser);
+        return newPost
     }
 
     async FindOneById(id: number): Promise<Posts>{
@@ -103,10 +123,37 @@ export class PostService {
         return {message: "Update successfully !"};
     }
 
-    async DeletePost(id: number):Promise<object|any>{
-        const action = await this.postRepository.delete({id});
-        if(action.affected === 0)
+    async DeleteUserPost(userId: number){
+        await this
+        await this.postRepository.delete({user: {id: userId}});
+    }
+
+    async DeletePost(id: number, userId: number):Promise<object|any>{
+        const post = await this.FindOneById(id);
+        if(!post)
             throw new NotFoundException("Post Not Found !");
+        if(post.user.id !== userId)
+            throw new ForbiddenException("Cannot delete other's post")
+
+        await this.postRepository.delete({id});
+        
+        const user = await this.userService.FindOne(undefined, undefined, post.user.id)
+        const updatedUser: UserDto = {
+            publishedPostCount: user.postPublishedCount -= 1,
+            firstName: user.firstName, 
+            lastName: user.lastName, 
+            phoneNum: user.phoneNum, 
+            password: user.password,
+            email: user.email, 
+            bio: user.bio,
+            profilePicturePath: user.profilePicturePath,
+            dateOfBirth: user.dateOfBirth,
+            username: user.username,
+            confirmPassword: user.password,
+            gender: user.gender,
+            roleId: user.role.id
+        }
+        await this.userService.UpdateUserInfor(user.id, updatedUser);
         return {message: "Deleted !"};
     }
 }
